scripts/python/opentelemetry-demo/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc